+++ visualization/app/codeCharta/services/loadFile/fileRoot.ts
@@ -1,0 +1,8 @@
+export const fileRoot = {
+	rootName: "root",
+	rootPath: `/root`,
+	updateRoot: (name: string) => {
+		fileRoot.rootName = name
+		fileRoot.rootPath = `/${name}`
+	}
+}