+++ visualization/app/codeCharta/ui/searchPanel/searchBar/utils/isPatternBlacklisted.spec.ts
@@ -1,20 +1,19 @@
-import { BlacklistType } from "../../../../codeCharta.model"
 import { isPatternBlacklisted } from "./isPatternBlacklisted"
 
 describe("isPatternBlacklisted", () => {
 	it("should recognized an exact match", () => {
-		expect(isPatternBlacklisted([{ type: BlacklistType.flatten, path: "*needle*" }], BlacklistType.flatten, "*needle*")).toBe(true)
+		expect(isPatternBlacklisted([{ type: "flatten", path: "*needle*" }], "flatten", "*needle*")).toBe(true)
 	})
 
 	it("should recognized * matches", () => {
-		expect(isPatternBlacklisted([{ type: BlacklistType.flatten, path: "*needle*" }], BlacklistType.flatten, "needle")).toBe(true)
+		expect(isPatternBlacklisted([{ type: "flatten", path: "*needle*" }], "flatten", "needle")).toBe(true)
 	})
 
 	it("should not recognized * vs absolute value", () => {
-		expect(isPatternBlacklisted([{ type: BlacklistType.flatten, path: "*needle*" }], BlacklistType.flatten, "/needle")).toBe(false)
+		expect(isPatternBlacklisted([{ type: "flatten", path: "*needle*" }], "flatten", "/needle")).toBe(false)
 	})
 
 	it("should ignore different blacklist type", () => {
-		expect(isPatternBlacklisted([{ type: BlacklistType.flatten, path: "*needle*" }], BlacklistType.exclude, "*needle*")).toBe(false)
+		expect(isPatternBlacklisted([{ type: "flatten", path: "*needle*" }], "exclude", "*needle*")).toBe(false)
 	})
 })