+++ visualization/app/codeCharta/ui/toolBar/toolBar.component.spec.ts
@@ -1,6 +1,5 @@
 import { TestBed } from "@angular/core/testing"
 import { render } from "@testing-library/angular"
-import { mocked } from "ts-jest/utils"
 import { LoadFileService } from "../../services/loadFile/loadFile.service"
 import { hoveredNodeIdSelector } from "../../state/store/appStatus/hoveredNodeId/hoveredNodeId.selector"
 import { ThreeCameraService } from "../codeMap/threeViewer/threeCamera.service"
@@ -12,7 +11,7 @@ import { ToolBarModule } from "./toolBar.module"
 jest.mock("../../state/store/appStatus/hoveredNodeId/hoveredNodeId.selector", () => ({
 	hoveredNodeIdSelector: jest.fn()
 }))
-const mockedHoveredNodeIdSelector = mocked(hoveredNodeIdSelector)
+const mockedHoveredNodeIdSelector = jest.mocked(hoveredNodeIdSelector)
 
 describe("ToolBarComponent", () => {
 	beforeEach(() => {