+++ visualization/app/codeCharta/util/aggregationGenerator.ts
@@ -1,8 +1,8 @@
 import { CodeMapNode, CCFile, NodeType } from "../codeCharta.model"
-import { LoadFileService } from "../services/loadFile/loadFile.service"
 import { FileNameHelper } from "./fileNameHelper"
 import { getUpdatedPath } from "./nodePathHelper"
 import packageJson from "../../../package.json"
+import { fileRoot } from "../services/loadFile/fileRoot"
 
 export class AggregationGenerator {
 	private static projectNameArray: string[] = []
@@ -36,11 +36,11 @@ export class AggregationGenerator {
 				exportedFileSize: this.fileSizesSum
 			},
 			map: {
-				name: LoadFileService.ROOT_NAME,
+				name: fileRoot.rootName,
 				type: NodeType.FOLDER,
 				children: [],
 				attributes: {},
-				path: LoadFileService.ROOT_PATH
+				path: fileRoot.rootPath
 			},
 			settings: {
 				fileSettings: {